+++ src/main/java/ru/liga/oldrussiantinderbot/service/UserService.java
@@ -7,9 +7,9 @@ import java.util.List;
 public interface UserService {
     List<User> getAllUsers();
 
-    User getUser(int id);
+    User getUser(Long id);
 
     void saveUser(User user);
 
-    void deleteUser(int id);
+    void deleteUser(Long id);
 }
