+++ src/main/java/ru/liga/oldrussiantinderbot/service/UserServiceImpl.java
@@ -18,7 +18,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User getUser(int id) {
+    public User getUser(Long id) {
         User user = null;
         Optional<User> byId = userRepository.findById(id);
         if (byId.isPresent()){
@@ -33,7 +33,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUser(int id) {
+    public void deleteUser(Long id) {
         userRepository.deleteById(id);
     }
 }
